1.更新ddl 2.更新dto 3.更新mock rest 4.更新pom

+++ src/main/java/com/mario1oreo/projects/business/pontus/dto/BiIncomeInfo.java
@@ -1,23 +1,26 @@
 package com.mario1oreo.projects.business.pontus.dto;
 
+
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class BiIncomeInfo {
 
-    private long id;
+    private int id;
     private String productId;
     private String batchNo;
-    private double productQuantity;
-    private double salePrice;
-    private double costPrice;
-    private double expressQuantity;
-    private double expressFee;
-    private double netProfit;
+    private BigDecimal productQuantity;
+    private BigDecimal salePrice;
+    private BigDecimal costPrice;
+    private BigDecimal expressQuantity;
+    private BigDecimal expressFee;
+    private BigDecimal netProfit;
     private LocalDate partition;
     private String status;
     private String createBy;
     private String updateBy;
+
 }
