+++ src/main/java/com/mario1oreo/projects/business/pontus/dto/BiInventoryInfo.java
@@ -1,22 +1,23 @@
 package com.mario1oreo.projects.business.pontus.dto;
 
-
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class BiInventoryInfo {
 
-    private long id;
+  private int id;
     private String productId;
     private String batchNo;
     private String productName;
     private String productUnit;
-    private double productQuantity;
+  private BigDecimal productQuantity;
     private LocalDate partition;
     private String status;
     private String createBy;
     private String updateBy;
+
 
 }
