+++ src/main/java/com/mario1oreo/projects/business/pontus/dto/PrdInventoryInfo.java
@@ -1,16 +1,19 @@
 package com.mario1oreo.projects.business.pontus.dto;
 
+
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 public class PrdInventoryInfo {
 
-    private long id;
+  private int id;
     private String productId;
     private String batchNo;
     private String productName;
     private String productUnit;
-    private double productQuantity;
+  private BigDecimal productQuantity;
     private String status;
     private String createBy;
     private String updateBy;
